requirements.txt
fastapi>=0.110
uvicorn>=0.27
PyYAML>=6.0
pydantic>=2.6
